Add vim-indentwise plugin
This offers indentation-based movement.

`[-` moves to previous line of lesser indent than the current line.
`[+` moves to previous line of greater indent than the current line.
`[=` moves to previous line of same indent as the current line that is
     separated from the current line by lines of different indents.

`]-` moves to next line of lesser indent than the current line.
`]+` moves to next line of greater indent than the current line.
`]=` moves to next line of same indent as the current line that is
     separated from the current line by lines of different indents.