Add Pyright to LSP sources
This offers faster type hints than mypy, but they're not always as good
because of the lack of type support.

It would be nice to get Blink to deduplicate answers from other LSPs,
but that's not yet supported.
See https://github.com/Saghen/
blink.cmp/issues/
1222

(Broken to avoid backlink.)